LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-autospace/text-autospace-first-line-001-expected.txt

国国XX国
国国XX国

FAIL text-autospace-first-line-001 assert_array_equals: expected property 1 to be 90 but got 80 (expected array [80, 90, 80, 90, 80, 40, 45, 40, 45, 40] got [80, 80, 80, 80, 80, 40, 40, 40, 40, 40])
